Fix typo which causes SD card status responses to fail
Removes delay times within SD Driver's internal SendCmd() calls.
Removed SD interface, placed virtual methods as part of SD class,
and separated CRC logic into utility file crc.hpp

Fix lint, tidy, and header inclusion issues for crc.hpp

Fix crc.hpp not including type_traits

Resolves #444
